Continuation monad transformer!
This one took quite a while to get right and I'm mightily pleased that
I managed to push it throught the type checker at all.

The key trick turned out to be to add an extra parameter to the data type
which tracks intermediate state. For a while I thought it was possible
to hide it using forall quantification but that turned out not to work
when control operators was involved.