Move `tell` and `listen` to the Writer class
The function `pass` is hard to move into the class because it
changes the category and that parameter is fixed in the class.
It is possible to make it another parameter to the monad instead
but that is really inflexible in practice and makes it hard to
make instances of the class. I'm going to leave `pass` out of
the class for now.